pyproject.toml
[tool.poetry]
name = "llm-wrapper"
version = "0.1.0"
description = ""
authors = ["Masatsugar <20656329+Masatsugar@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
openai = "^1.58.1"
fire = "^0.7.0"
tenacity = "^9.0.0"
pillow = "^11.0.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
